refactor(step2):  pakage
pakage 분리
