+++ src/main/java/lotto/controller/LottoRun.java
@@ -1,0 +1,4 @@
+package lotto.controller;
+
+public class LottoRun {
+}
