+++ src/main/java/lotto/domain/LottoInspector.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.HashMap;
 import java.util.List;
