+++ src/main/java/lotto/domain/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.generator.LottoGenerator;
 import lotto.generator.RandomNumberGenerator;
