+++ src/main/java/lotto/view/LottoResultView.java
@@ -1,0 +1,4 @@
+package lotto.view;
+
+public class LottoResultView {
+}
