+++ src/test/java/lotto/LottoMachineTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.LottoMachine;
+import lotto.domain.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
