+++ src/test/java/lotto/LottoNumberTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
